+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- Add method `hashAdvanced` to `Argon2Advanced` which lets the caller specify the argon2 version
+- Add methods `generateSalt` to `Argon2Advanced` which generates salt with either the default or the given length
+
+### Added
 - Improved the [compatibility list](compatibility-tests/README.md)
 
 ## [2.7] - 2020-04-15